provide tooltip for oveflow text and uncommon icons

+++ src/utils/getDayName.ts
@@ -1,6 +1,6 @@
 const getDayName = (
   data: string = "2000-01-27",
-  weekdayFormat: "short" | "long" | "narrow" = "long",
+  weekdayFormat: "short" | "long" | "narrow" = "short",
 ) => {
   return new Date(data).toLocaleDateString("en-US", { weekday: weekdayFormat });
 };
